+++ src/config.rs
@@ -1,20 +1,29 @@
-use std::{env, fmt};
-
+#[cfg(not(feature = "blocking"))]
+use governor::{
+    clock::{QuantaClock, QuantaInstant},
+    middleware::NoOpMiddleware,
+    state::{InMemoryState, NotKeyed},
+    Quota, RateLimiter,
+};
 #[cfg(feature = "blocking")]
 use reqwest::blocking::{Client, RequestBuilder, Response};
 use reqwest::header::USER_AGENT;
 #[cfg(not(feature = "blocking"))]
 use reqwest::{Client, RequestBuilder, Response};
 use reqwest::{Method, Url};
+use std::{env, fmt};
+#[cfg(not(feature = "blocking"))]
+use std::{num::NonZeroU32, sync::Arc, time::Duration};
 
-use crate::types::ErrorResponse;
-use crate::{Error, Result};
+use crate::{error::types::ErrorResponse, Error, Result};
 
 pub struct Config {
     pub(crate) user_agent: String,
     pub(crate) api_key: String,
     pub(crate) base_url: Url,
     pub(crate) client: Client,
+    #[cfg(not(feature = "blocking"))]
+    limiter: Arc<RateLimiter<NotKeyed, InMemoryState, QuantaClock, NoOpMiddleware<QuantaInstant>>>,
 }
 
 impl Config {
@@ -27,15 +36,33 @@ impl Config {
             )
             .expect("env variable `RESEND_BASE_URL` should be a valid URL");
 
-        let env_user_agent = env::var("RESEND_USER_AGENT").unwrap_or_else(|_| {
-            format!("{}/{}", env!("CARGO_PKG_NAME"), env!("CARGO_PKG_VERSION"))
-        });
+        let env_user_agent = format!("{}/{}", env!("CARGO_PKG_NAME"), env!("CARGO_PKG_VERSION"));
+
+        // ==== Rate limiting is a non-blocking thing only ====
+        #[cfg(not(feature = "blocking"))]
+        let rate_limit_per_sec = env::var("RATE_LIMIT")
+            .unwrap_or_else(|_| "10".to_owned())
+            .parse::<u32>()
+            .expect("env variable `RATE_LIMIT` should be a valid u32");
+
+        #[cfg(not(feature = "blocking"))]
+        let quota = Quota::with_period(Duration::from_millis(1100))
+            .expect("Valid quota")
+            .allow_burst(
+                NonZeroU32::new(rate_limit_per_sec).expect("Rate limit is a valid non zero u32"),
+            );
+
+        #[cfg(not(feature = "blocking"))]
+        let limiter = Arc::new(RateLimiter::direct(quota));
+        // ====================================================
 
         Self {
             user_agent: env_user_agent,
             api_key: api_key.to_owned(),
             base_url: env_base_url,
             client,
+            #[cfg(not(feature = "blocking"))]
+            limiter,
         }
     }
 
@@ -54,7 +81,15 @@ impl Config {
 
     #[maybe_async::maybe_async]
     pub async fn send(&self, request: RequestBuilder) -> Result<Response> {
+        #[cfg(not(feature = "blocking"))]
+        {
+            let jitter =
+                governor::Jitter::new(Duration::from_millis(10), Duration::from_millis(50));
+            self.limiter.until_ready_with_jitter(jitter).await;
+        }
+
         let request = request.build()?;
+
         let response = self.client.execute(request).await?;
 
         match response.status() {
@@ -75,178 +110,5 @@ impl fmt::Debug for Config {
             .field("user_agent", &self.user_agent.as_str())
             .field("base_url", &self.base_url.as_str())
             .finish_non_exhaustive()
-    }
-}
-
-pub mod types {
-    use serde::Deserialize;
-
-    /// Error returned as a response.
-    ///
-    /// <https://resend.com/docs/api-reference/errors>
-    #[derive(Debug, Clone, Deserialize, thiserror::Error)]
-    #[error("resend error: {message}")]
-    pub struct ErrorResponse {
-        #[serde(rename = "statusCode")]
-        pub status_code: u16,
-        pub message: String,
-        pub name: String,
-    }
-
-    impl ErrorResponse {
-        /// Returns the [`ErrorKind`].
-        #[must_use]
-        pub fn kind(&self) -> ErrorKind {
-            ErrorKind::from(self.name.as_str())
-        }
-    }
-
-    /// Error type for operations of a [`Client`].
-    ///
-    /// <https://resend.com/docs/api-reference/errors>
-    ///
-    /// [`Client`]: crate::Client
-    #[non_exhaustive]
-    #[derive(Debug, Copy, Clone)]
-    pub enum ErrorKind {
-        /// Error name is not in the API spec.
-        Unrecognized,
-
-        /// 401 Unauthorized.
-        ///
-        /// - `missing_api_key`
-        ///
-        /// Missing API key in the authorization header.
-        ///
-        /// Include the following header `Authorization: Bearer YOUR_API_KEY` in the request.
-        MissingApiKey,
-
-        /// 403 Forbidden.
-        ///
-        /// - `invalid_api_key`
-        ///
-        /// The API key is not valid.
-        ///
-        /// Generate a new API key in the dashboard.
-        InvalidApiKey,
-
-        /// 403 Forbidden.
-        ///
-        /// - `invalid_from_address`
-        ///
-        /// The from address is not valid.
-        ///
-        /// Review your existing domains in the dashboard.
-        InvalidFromAddress,
-
-        /// 403 Forbidden.
-        ///
-        /// - `invalid_to_address`
-        ///
-        /// You can only send testing emails to your own email address.
-        ///
-        /// In order to send emails to any external address, you need to add a domain and
-        /// use that as the `from` address instead of `onboarding@resend.dev`.
-        InvalidToAddress,
-
-        /// 404 Not Found.
-        ///
-        /// - `not_found`
-        ///
-        /// The requested endpoint does not exist.
-        ///
-        /// Change your request URL to match a valid API endpoint.
-        NotFound,
-
-        /// 405 Method Not Allowed.
-        ///
-        /// - `method_not_allowed`
-        ///
-        /// This endpoint does not support the specified HTTP method.
-        ///
-        /// Change the HTTP method to follow the documentation for the endpoint.
-        MethodNotAllowed,
-
-        /// 422 Unprocessable Content.
-        ///
-        /// - `missing_required_field`
-        ///
-        /// The request body is missing one or more required fields.
-        /// Check the error message to see the list of missing fields.
-        MissingRequiredField,
-
-        /// 422 Unprocessable Content.
-        ///
-        /// - `invalid_attachment`
-        ///
-        /// Attachment must have either a `content` or `path`.
-        ///
-        /// Attachments must either have a `content` (strings, Buffer, or Stream contents) or
-        /// `path` to a remote resource (better for larger attachments).
-        InvalidAttachment,
-
-        /// 422 Unprocessable Content.
-        ///
-        /// - `invalid_scope`
-        ///
-        /// This endpoint does not support the specified scope.
-        /// Change the scope to follow the documentation for the endpoint.
-        InvalidScope,
-
-        /// 429 Too Many Requests.
-        ///
-        /// - `rate_limit_exceeded`
-        ///
-        ///  Too many requests. Please limit the number of requests per second.
-        /// Or contact support to increase rate limit.
-        ///
-        /// You should read the response headers and reduce the rate at which you request the API.
-        /// This can be done by introducing a queue mechanism or reducing the number of
-        /// concurrent requests per second.
-        ///
-        /// If you have specific requirements, contact support to request a rate increase.
-        RateLimitExceeded,
-
-        /// 429 Too Many Requests.
-        ///
-        /// - `daily_quota_exceeded`
-        ///
-        /// You have reached your daily email sending quota.
-        ///
-        /// Upgrade your plan to remove daily quota limit or
-        /// wait until 24 hours have passed to continue sending.
-        DailyQuotaExceeded,
-
-        /// 500 Internal Server Error.
-        ///
-        /// - `internal_server_error`
-        ///
-        /// An unexpected error occurred.
-        ///
-        /// Try the request again later. If the error does not resolve,
-        /// check our [`status page`] for service updates.
-        ///
-        /// [`status page`]: https://resend-status.com/
-        InternalServerError,
-    }
-
-    impl<T: AsRef<str>> From<T> for ErrorKind {
-        fn from(value: T) -> Self {
-            match value.as_ref() {
-                "missing_api_key" => Self::MissingApiKey,
-                "invalid_api_key" => Self::InvalidApiKey,
-                "invalid_from_address" => Self::InvalidFromAddress,
-                "invalid_to_address" => Self::InvalidToAddress,
-                "not_found" => Self::NotFound,
-                "method_not_allowed" => Self::MethodNotAllowed,
-                "missing_required_field" => Self::MissingRequiredField,
-                "invalid_attachment" => Self::InvalidAttachment,
-                "invalid_scope" => Self::InvalidScope,
-                "rate_limit_exceeded" => Self::RateLimitExceeded,
-                "daily_quota_exceeded" => Self::DailyQuotaExceeded,
-                "internal_server_error" => Self::InternalServerError,
-                _ => Self::Unrecognized,
-            }
-        }
     }
 }